refactor: application 요청함수 단계별 파일 분리
