+++ src/apis/admin/application/status/index.ts
@@ -1,0 +1,11 @@
+import instance from 'apis/instance';
+import { DailyWorkTimeData } from 'apis/schedule/getDailyWorkers';
+import { AxiosResponse } from 'axios';
+
+export const getApplyStatus = (params: { startWeekDate: string }): Promise<AxiosResponse<GetApplyStatusResponse>> => {
+  return instance.get(`/schedule/remain/week`, { params });
+};
+
+interface GetApplyStatusResponse {
+  applyStatus: DailyWorkTimeData[][];
+}
